fix: Merge colors
