+++ src/layout/MobileCalculatorPage.jsx
@@ -15,7 +15,7 @@ import { ParameterSearch } from "../pages/PolicyPage.jsx";
 import PolicyRightSidebar from "../pages/policy/PolicyRightSidebar.jsx";
 
 import style from "../style";
-import colors from "../redesign/style/colors";
+import colors from "../style/colors";
 import spacing from "../style/spacing";
 
 /**
