+++ src/pages/policy/output/LaborSupplyResponse.jsx
@@ -41,7 +41,7 @@ function ImpactPlot(props) {
             },
           },
           connector: {
-            line: { color: style.colors.GRAY, width: 1, dash: "dot" },
+            line: { color: style.colors.MEDIUM_LIGHT_GRAY, width: 1, dash: "dot" },
           },
           hovertemplate: `<b>%{x}</b><extra></extra>`,
         },
