--- src/redesign/style/colors.jsx
@@ -1,37 +1,0 @@
-const BLUE_LIGHT = "#D8E6F3";
-const BLUE_PRIMARY = "#2C6496";
-const BLUE_PRESSED = "#17354F";
-const BLUE_98 = "#F7FAFD";
-const TEAL_LIGHT = "#D7F4F2";
-const TEAL_ACCENT = "#39C6C0";
-const TEAL_PRESSED = "#40e0d0";
-const DARKEST_BLUE = "#0C1A27";
-const DARK_GRAY = "#616161";
-const GRAY = "#808080";
-const LIGHT_GRAY = "#F2F2F2";
-const MEDIUM_DARK_GRAY = "#D2D2D2";
-const WHITE = "#FFFFFF";
-const TEAL_98 = "#F7FDFC";
-const BLACK = "#000000";
-const DARK_BLUE_HOVER = "#1d3e5e";
-
-const colors = {
-  BLUE_LIGHT,
-  BLUE_PRIMARY,
-  BLUE_PRESSED,
-  BLUE_98,
-  TEAL_LIGHT,
-  TEAL_ACCENT,
-  TEAL_PRESSED,
-  DARKEST_BLUE,
-  DARK_GRAY,
-  GRAY,
-  LIGHT_GRAY,
-  MEDIUM_DARK_GRAY,
-  WHITE,
-  TEAL_98,
-  BLACK,
-  DARK_BLUE_HOVER,
-};
-
-export default colors;
