--- src/style/colors.jsx
@@ -1,39 +1,0 @@
-const WHITE = "#FFF";
-const BLUE = "#2C6496";
-const GRAY = "#BDBDBD";
-const MEDIUM_DARK_GRAY = "#D2D2D2";
-const DARK_GRAY = "#616161";
-const GREEN = "#29d40f";
-const LIGHT_GRAY = "#F2F2F2";
-const DARK_RED = "#b50d0d";
-const BLACK = "#000";
-const TEAL_ACCENT = "#39C6C0";
-const TEAL_LIGHT = "#F7FDFC";
-const TEAL_PRESSED = "#227773";
-const BLUE_98 = "#D8E6F3";
-const DARKEST_BLUE = "#0C1A27";
-const DARK_BLUE_HOVER = "#1d3e5e";
-const BLUE_LIGHT = "#D8E6F3";
-const BLUE_PRESSED = "#17354F";
-
-const colors = {
-  WHITE,
-  BLUE,
-  GRAY,
-  DARK_GRAY,
-  LIGHT_GRAY,
-  GREEN,
-  DARK_RED,
-  BLACK,
-  MEDIUM_DARK_GRAY,
-  TEAL_ACCENT,
-  TEAL_PRESSED,
-  TEAL_LIGHT,
-  BLUE_98,
-  DARKEST_BLUE,
-  DARK_BLUE_HOVER,
-  BLUE_LIGHT,
-  BLUE_PRESSED,
-};
-
-export default colors;
